Refactor generation functions into project_utils
